+++ src/sdk/modules/smartSessionsValidator/Types.ts
@@ -3,6 +3,7 @@ import type {
   SmartSessionMode
 } from "@rhinestone/module-sdk"
 import type { AbiFunction, Address, Hex, OneOf } from "viem"
+import type { KeyGenData } from "../../clients/decorators/dan/decorators/keyGen"
 import type { AnyReferenceValue } from "../utils/Helpers"
 import type { Execution } from "../utils/Types"
 
@@ -54,11 +55,15 @@ export type SmartSessionModeType =
  */
 export type UsePermissionModuleData = {
   /** The permission ID for the session. */
-  permissionId: Hex
+  permissionIds: Hex[]
   /** The mode of the smart session. */
   mode?: SmartSessionModeType
   /** Data for enabling the session. */
   enableSessionData?: EnableSessionData
+  /** Key generation data for the session. */
+  keyGenData?: KeyGenData
+  /** The index of the permission ID to use for the session. Defaults to 0. */
+  permissionIdIndex?: number
 }
 
 type OptionalSessionKeyData = OneOf<