+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @biconomy/sdk
 
+## 0.0.10
+
+### Patch Changes
+
+- Added Distributed Session Keys w/ Ownable & Session examples
+
+## 0.0.9
+
+### Patch Changes
+
+- Added DAN helpers, keyGen + sigGen
+
 ## 0.0.8
 
 ### Patch Changes